+++ .versionrc.js
@@ -1,0 +1,5 @@
+const { VersionConfigTypes } = require('./src/constants')
+
+module.exports = {
+  types: VersionConfigTypes,
+}