+++ src/Prose.tsx
@@ -1,0 +1,26 @@
+import React, { useContext } from 'react'
+import { twMerge } from 'tailwind-merge'
+import { ThemeContext } from './ThemeProvider'
+
+export interface ProseProps {
+  className?: string
+  children: React.ReactNode
+}
+
+export function Prose({ className, children }: ProseProps) {
+  const theme = useContext(ThemeContext)
+
+  return (
+    <div
+      className={twMerge(
+        'prose prose-headings:font-thesans prose-headings:font-bold prose-h1:text-2xl prose-h2:text-xl prose-h3:text-lg prose-h4:text-md',
+        theme.primaryProseHover,
+        className
+      )}
+    >
+      {children}
+    </div>
+  )
+}
+
+export default Prose