+++ src/ThemeProvider.tsx
@@ -1,4 +1,4 @@
-import { createContext } from 'react'
+import React, { createContext } from 'react'
 
 type Theme = {
   primaryBg: string
@@ -7,6 +7,9 @@ type Theme = {
   primaryBorderHover: string
   primaryText: string
   primaryTextHover: string
+  primaryProseHover: string
+  primaryFill: string
+  primaryFillHover: string
 }
 
 interface Props {
@@ -21,11 +24,14 @@ const BASE_THEME: Theme = {
   primaryBorderHover: 'hover:border-uzh-red-40',
   primaryText: 'text-uzh-red-100',
   primaryTextHover: 'hover:text-uzh-red-100',
+  primaryProseHover: 'hover:prose-a:text-uzh-red-100',
+  primaryFill: 'fill-uzh-red-80',
+  primaryFillHover: 'hover:fill-uzh-red-100',
 }
 
 export const ThemeContext = createContext<Theme>(BASE_THEME)
 
-function ThemeProvider({ theme, children }: Props) {
+export function ThemeProvider({ theme, children }: Props) {
   return (
     <ThemeContext.Provider value={{ ...BASE_THEME, ...theme }}>
       {children}