+++ src/constants.js
@@ -61,6 +61,65 @@ const PrettierConfig = {
   trailingComma: 'es5',
 }
 
+const VersionConfigTypes = [
+  {
+    type: 'feat',
+    section: 'Features',
+  },
+  {
+    type: 'enhance',
+    section: 'Enhancements',
+  },
+  {
+    type: 'fix',
+    section: 'Bug Fixes',
+  },
+  {
+    type: 'docs',
+    section: 'Documentation',
+  },
+  {
+    type: 'refactor',
+    section: 'Refactors',
+  },
+  {
+    type: 'perf',
+    section: 'Performance',
+  },
+  {
+    type: 'deploy',
+    section: 'Deployment',
+  },
+  {
+    type: 'deps',
+    section: 'Dependencies',
+  },
+  {
+    type: 'build',
+    section: 'Build and CI',
+  },
+  {
+    type: 'ci',
+    section: 'Build and CI',
+  },
+  {
+    type: 'chore',
+    section: 'Other',
+  },
+  {
+    type: 'wip',
+    section: 'Other',
+  },
+  {
+    type: 'test',
+    section: 'Other',
+  },
+  {
+    type: 'style',
+    section: 'Other',
+  },
+]
+
 module.exports = {
   TailwindColorsUZH,
   TailwindFonts,
@@ -68,4 +127,5 @@ module.exports = {
   TailwindProdPlugins,
   ESLintConfig,
   PrettierConfig,
+  VersionConfigTypes,
 }